+++ test/binary_parser.js
@@ -12,7 +12,7 @@ for (var i = 0; i < 64; i++) {
 
 function BinaryParser (bigEndian, allowExceptions) {
   if(!(this instanceof BinaryParser)) return new BinaryParser(bigEndian, allowExceptions);
-  
+
 	this.bigEndian = bigEndian;
 	this.allowExceptions = allowExceptions;
 };
@@ -48,7 +48,7 @@ BinaryParser.decodeInt = function decodeInt (data, bits, signed, forceBigEndian)
   var b = new this.Buffer(this.bigEndian || forceBigEndian, data)
       , x = b.readBits(0, bits)
       , max = maxBits[bits]; //max = Math.pow( 2, bits );
-  
+
   return signed && x >= max / 2
       ? x - max
       : x;
@@ -271,7 +271,7 @@ BinaryParser.hprint = function hprint (s) {
     if (s.charCodeAt(i) < 32) {
       number = s.charCodeAt(i) <= 15
         ? "0" + s.charCodeAt(i).toString(16)
-        : s.charCodeAt(i).toString(16);        
+        : s.charCodeAt(i).toString(16);
       process.stdout.write(number + " ")
     } else {
       number = s.charCodeAt(i) <= 15
@@ -280,7 +280,7 @@ BinaryParser.hprint = function hprint (s) {
         process.stdout.write(number + " ")
     }
   }
-  
+
   process.stdout.write("\n\n");
 };